bee-engine/requirements.txt
mysqlclient==1.3.7
pyserial==3.4
